requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colored==1.4.4
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
importlib-metadata==6.8.0
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.3.0
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
